+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project, at least loosely, adheres to [Semantic Versioning](https://sem
 ## Unreleased
 ### Changed
 - global clock files now emit a warning instead of an exception if expired and the download fails
+### Added
+### Fixed
+- dmxparse outputs to dmxparse.out if save=True
 
 ## [0.9.1] 2022-08-12
 ### Changed